requirements.txt
openai 
langchain
langchain-openai
langchain_pinecone
langsmith
pydantic
pinecone-client
streamlit
fastapi
uvicorn
pypdf
python-dotenv
tavily-python
langchain-community
tavily-python
langchainhub
duckduckgo-search
beautifulsoup4
python-dotenv
faiss-cpu
duckduckgo_search
speechrecognition
PyAudio
py3-tts
pydantic